Fix zombie `convert` process left behind when thumbnail generation times out (#44378).
Patch by Go MAEDA (user:maeda).


git-svn-id: https://svn.redmine.org/redmine/trunk@25033 e93f8b46-1217-0410-a6f0-8f06a7374b81

+++ lib/redmine/thumbnail.rb
@@ -66,7 +66,10 @@ module Redmine
             end
           end
         rescue Timeout::Error
-          Process.kill('KILL', pid)
+          if pid
+            Process.kill('KILL', pid)
+            Process.detach(pid)
+          end
           logger.error("Creating thumbnail timed out:\nCommand: #{cmd}")
           return nil
         end
